Floating assistant, external media tracking, widget, and in-app playe…

+++ app/src/main/java/com/example/mediavault/DailyProgress.java
@@ -1,11 +1,11 @@
 package com.example.mediavault;
 
 public class DailyProgress {
-    public int pagesRead;
-    public int episodesWatched;
-    public int minutesWatched;
+    public float pagesRead;
+    public float episodesWatched;
+    public float minutesWatched;
 
-    public DailyProgress(int pages, int episodes, int minutes) {
+    public DailyProgress(float pages, float episodes, float minutes) {
         this.pagesRead = pages;
         this.episodesWatched = episodes;
         this.minutesWatched = minutes;
